+++ foc_core/include/LS_Config/foc_cfg_filter.h
@@ -45,7 +45,7 @@
 #define FOC_FILTER_SENSOR_ANGLE_LPF_ALPHA             0.80f
 
 /* ── ENCODER_SPEED ── */
-#define FOC_ENCODER_SPEED_WINDOW_SIZE                 32U
+#define FOC_ENCODER_SPEED_WINDOW_SIZE                 8U
 /* Kalman parameters (used when type == KALMAN) */
 #define FOC_FILTER_ENCODER_SPEED_KALMAN_MEAS_ERR      0.5f
 #define FOC_FILTER_ENCODER_SPEED_KALMAN_EST_ERR       0.0f
@@ -56,14 +56,12 @@
 
 /* ── SMO_SPEED ── */
 #define FOC_FILTER_SMO_SPEED           FOC_FILTER_TYPE_LPF1
-#define FOC_SMO_ANGLE_HISTORY_SIZE                     8U
-#define FOC_SMO_SPEED_WINDOW_SIZE                      8U
 /* Kalman parameters (used when type == KALMAN) */
 #define FOC_FILTER_SMO_SPEED_KALMAN_MEAS_ERR           0.5f
 #define FOC_FILTER_SMO_SPEED_KALMAN_EST_ERR            0.0f
 #define FOC_FILTER_SMO_SPEED_KALMAN_PROC_NOISE         0.01f
 #define FOC_FILTER_SMO_SPEED_KALMAN_INIT               0.0f
 /* LPF1 parameter (used when type == LPF1) */
-#define FOC_FILTER_SMO_SPEED_LPF_ALPHA                 0.4f
+#define FOC_FILTER_SMO_SPEED_LPF_ALPHA                 0.2f
 
 #endif /* FOC_CFG_FILTER_H */